---
THE ARCHITECT ᛫ Sebastien Rousseau ᛫ https://sebastienrousseau.com
THE ENGINE ᛞ EUXIS ᛫ Enterprise Unified Execution Intelligence System ᛫ https://euxis.co
# Conflicts:
#	apps/cli/src/terminal.cpp

+++ apps/etx/src/widgets/bento_layout.cpp
@@ -35,7 +35,7 @@ int BentoLayout::column_count() const {
 }
 
 int BentoLayout::card_count() const {
-    return cards_.size();
+    return static_cast<int>(cards_.size());
 }
 
 void BentoLayout::resizeEvent(QResizeEvent* event) {
